root: Try to preserve relative positions of floating containers
This makes the behavior of floating containers more consistent with i3.
The workspace-relative coordinates of the container are scaled when the
size of the workspace it is on changes or when the container is moved
between workspaces.

For scratchpad containers, add a new state that preserves the dimensions
of the last workspace the window appeared on. This is necessary because
after a container is hidden in the scratchpad, we expect it to be in the
same relative position when it reappears. We can't just use the
container's attached workspace because that workspace's dimensions might
have been changed or the workspace as a whole could have been destroyed.